scrapbox: split files
