+++ index.js
@@ -31,6 +31,7 @@ const allBots = [
 	// ...(word2vecInstalled ? ['vocabwar'] : []),
 	'ricochet-robots',
 	'scrapbox',
+	'scrapbox/mute',
 	'slack-log',
 	'welcome',
 	'deploy',
